Revert "organized spritesheet playing"
This reverts commit f9edfc08ec8a2a5517a2d765a1665fa1a074994a.

+++ Smash/LifeSpan.cpp
@@ -16,7 +16,7 @@ bool LifeSpan::hasEnded() {
 }
 
 void LifeSpan::reset(int life) {
-	life_ends_at = (int)gameClock.getElapsedTime().asMilliseconds() + (int)sf::milliseconds(life).asMilliseconds();
+	life_ends_at = (int)gameClock.getElapsedTime().asMilliseconds() + sf::milliseconds(life).asMilliseconds();
 }
 
 int LifeSpan::getLifeEnds() {
